[svn-r15462] Description:     Correct compiler warnings from Visual Studio.
Tested on:
    Mac OS X/32 10.5.4 (amazon) w/FORTRAN & C++
    (Too minor to require full h5committest)

+++ c++/src/H5Attribute.cpp
@@ -153,7 +153,7 @@ void Attribute::read( const DataType& mem_type, H5std_string& strg ) const
    {
       throw AttributeIException("Attribute::read", "Unable to get attribute size before reading");
    }
-   char* strg_C = new char [attr_size+1];
+   char* strg_C = new char [(size_t)attr_size+1];
    if (strg_C == NULL)
    {
       throw AttributeIException("Attribute::read", "Unable to allocate buffer to read the attribute");
